Move `StacksMicroblockHeader` and a few other things
