+++ stackslib/src/chainstate/stacks/block.rs
@@ -17,7 +17,6 @@
 use std::collections::HashMap;
 use std::io::{Read, Write};
 
-use sha2::{Digest, Sha512_256};
 use stacks_common::codec::{
     read_next, read_next_at_most, write_next, Error as codec_error, StacksMessageCodec,
     MAX_MESSAGE_LEN,
@@ -27,10 +26,10 @@ use stacks_common::types::chainstate::{
 };
 use stacks_common::util::hash::{MerkleTree, Sha512Trunc256Sum};
 use stacks_common::util::retry::BoundReader;
+#[cfg(test)]
 use stacks_common::util::secp256k1::MessageSignature;
 use stacks_common::util::vrf::*;
 
-use crate::burnchains::PrivateKey;
 use crate::chainstate::burn::operations::*;
 use crate::chainstate::burn::{ConsensusHash, *};
 use crate::chainstate::stacks::db::StacksBlockHeaderTypes;
@@ -677,163 +676,6 @@ impl StacksBlock {
     }
 }
 
-impl StacksMessageCodec for StacksMicroblockHeader {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        self.serialize(fd, false)
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<StacksMicroblockHeader, codec_error> {
-        let version: u8 = read_next(fd)?;
-        let sequence: u16 = read_next(fd)?;
-        let prev_block: BlockHeaderHash = read_next(fd)?;
-        let tx_merkle_root: Sha512Trunc256Sum = read_next(fd)?;
-        let signature: MessageSignature = read_next(fd)?;
-
-        // signature must be well-formed
-        // in tests, we sometimes use invalid signatures
-        #[cfg(not(any(test, feature = "testing")))]
-        let _ = signature
-            .to_secp256k1_recoverable()
-            .ok_or(codec_error::DeserializeError(
-                "Failed to parse signature".to_string(),
-            ))?;
-
-        Ok(StacksMicroblockHeader {
-            version,
-            sequence,
-            prev_block,
-            tx_merkle_root,
-            signature,
-        })
-    }
-}
-
-impl StacksMicroblockHeader {
-    pub fn sign(&mut self, privk: &StacksPrivateKey) -> Result<(), net_error> {
-        self.signature = MessageSignature::empty();
-        let mut bytes = vec![];
-        self.consensus_serialize(&mut bytes)
-            .expect("BUG: failed to serialize to a vec");
-
-        let mut digest_bits = [0u8; 32];
-        let mut sha2 = Sha512_256::new();
-
-        sha2.update(&bytes[..]);
-        digest_bits.copy_from_slice(sha2.finalize().as_slice());
-
-        let sig = privk
-            .sign(&digest_bits)
-            .map_err(|se| net_error::SigningError(se.to_string()))?;
-
-        self.signature = sig;
-        Ok(())
-    }
-
-    fn serialize<W: Write>(&self, fd: &mut W, empty_sig: bool) -> Result<(), codec_error> {
-        write_next(fd, &self.version)?;
-        write_next(fd, &self.sequence)?;
-        write_next(fd, &self.prev_block)?;
-        write_next(fd, &self.tx_merkle_root)?;
-        if empty_sig {
-            write_next(fd, &MessageSignature::empty())?;
-        } else {
-            write_next(fd, &self.signature)?;
-        }
-        Ok(())
-    }
-
-    pub fn check_recover_pubkey(&self) -> Result<Hash160, net_error> {
-        let mut digest_bits = [0u8; 32];
-        let mut sha2 = Sha512_256::new();
-
-        self.serialize(&mut sha2, true)
-            .expect("BUG: failed to serialize to a vec");
-        digest_bits.copy_from_slice(sha2.finalize().as_slice());
-
-        let mut pubk =
-            StacksPublicKey::recover_to_pubkey(&digest_bits, &self.signature).map_err(|_ve| {
-                test_debug!(
-                    "Failed to verify signature: failed to recover public key from {:?}: {:?}",
-                    &self.signature,
-                    &_ve
-                );
-                net_error::VerifyingError(
-                    "Failed to verify signature: failed to recover public key".to_string(),
-                )
-            })?;
-
-        pubk.set_compressed(true);
-        Ok(StacksBlockHeader::pubkey_hash(&pubk))
-    }
-
-    pub fn verify(&self, pubk_hash: &Hash160) -> Result<(), net_error> {
-        let pubkh = self.check_recover_pubkey()?;
-
-        if pubkh != *pubk_hash {
-            test_debug!(
-                "Failed to verify signature: public key did not recover to hash {}",
-                &pubkh.to_hex()
-            );
-            return Err(net_error::VerifyingError(format!(
-                "Failed to verify signature: public key did not recover to expected hash {}",
-                pubkh.to_hex()
-            )));
-        }
-
-        Ok(())
-    }
-
-    pub fn block_hash(&self) -> BlockHeaderHash {
-        let mut bytes = vec![];
-        self.consensus_serialize(&mut bytes)
-            .expect("BUG: failed to serialize to a vec");
-        BlockHeaderHash::from_serialized_header(&bytes[..])
-    }
-
-    /// Create the first microblock header in a microblock stream.
-    /// The header will not be signed
-    pub fn first_unsigned(
-        parent_block_hash: &BlockHeaderHash,
-        tx_merkle_root: &Sha512Trunc256Sum,
-    ) -> StacksMicroblockHeader {
-        StacksMicroblockHeader {
-            version: 0,
-            sequence: 0,
-            prev_block: parent_block_hash.clone(),
-            tx_merkle_root: tx_merkle_root.clone(),
-            signature: MessageSignature::empty(),
-        }
-    }
-
-    /// Create the first microblock header in a microblock stream for an empty microblock stream.
-    /// The header will not be signed
-    pub fn first_empty_unsigned(parent_block_hash: &BlockHeaderHash) -> StacksMicroblockHeader {
-        StacksMicroblockHeader::first_unsigned(parent_block_hash, &Sha512Trunc256Sum([0u8; 32]))
-    }
-
-    /// Create an unsigned microblock header from its parent
-    /// Return an error on overflow
-    pub fn from_parent_unsigned(
-        parent_header: &StacksMicroblockHeader,
-        tx_merkle_root: &Sha512Trunc256Sum,
-    ) -> Option<StacksMicroblockHeader> {
-        let next_sequence = match parent_header.sequence.checked_add(1) {
-            Some(next) => next,
-            None => {
-                return None;
-            }
-        };
-
-        Some(StacksMicroblockHeader {
-            version: 0,
-            sequence: next_sequence,
-            prev_block: parent_header.block_hash(),
-            tx_merkle_root: tx_merkle_root.clone(),
-            signature: MessageSignature::empty(),
-        })
-    }
-}
-
 impl StacksMessageCodec for StacksMicroblock {
     fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
         write_next(fd, &self.header)?;
@@ -924,11 +766,11 @@ impl StacksMicroblock {
     }
 
     pub fn sign(&mut self, privk: &StacksPrivateKey) -> Result<(), net_error> {
-        self.header.sign(privk)
+        self.header.sign(privk).map_err(Into::into)
     }
 
     pub fn verify(&mut self, pubk_hash: &Hash160) -> Result<(), net_error> {
-        self.header.verify(pubk_hash)
+        self.header.verify(pubk_hash).map_err(Into::into)
     }
 
     pub fn block_hash(&self) -> BlockHeaderHash {
