+++ stackslib/src/chainstate/stacks/transaction.rs
@@ -17,8 +17,9 @@
 use std::hash::{Hash, Hasher};
 use std::io::{Read, Write};
 
-use clarity::vm::representations::{ClarityName, ContractName};
-use clarity::vm::types::{QualifiedContractIdentifier, StandardPrincipalData};
+#[cfg(test)]
+use clarity::vm::representations::ClarityName;
+use clarity::vm::representations::ContractName;
 use clarity::vm::{ClarityVersion, Value};
 use stacks_common::codec::{read_next, write_next, Error as codec_error, StacksMessageCodec};
 use stacks_common::types::chainstate::StacksAddress;
@@ -29,81 +30,6 @@ use crate::burnchains::Txid;
 use crate::chainstate::stacks::{TransactionPayloadID, *};
 use crate::net::Error as net_error;
 use crate::util_lib::boot::boot_code_addr;
-
-impl StacksMessageCodec for TransactionContractCall {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        write_next(fd, &self.address)?;
-        write_next(fd, &self.contract_name)?;
-        write_next(fd, &self.function_name)?;
-        write_next(fd, &self.function_args)?;
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<TransactionContractCall, codec_error> {
-        let address: StacksAddress = read_next(fd)?;
-        let contract_name: ContractName = read_next(fd)?;
-        let function_name: ClarityName = read_next(fd)?;
-        let function_args: Vec<Value> = {
-            let mut bound_read = BoundReader::from_reader(fd, u64::from(MAX_TRANSACTION_LEN));
-            read_next(&mut bound_read)
-        }?;
-
-        // function name must be valid Clarity variable
-        if !StacksString::from(function_name.clone()).is_clarity_variable() {
-            warn!("Invalid function name -- not a clarity variable");
-            return Err(codec_error::DeserializeError(
-                "Failed to parse transaction: invalid function name -- not a Clarity variable"
-                    .to_string(),
-            ));
-        }
-
-        Ok(TransactionContractCall {
-            address,
-            contract_name,
-            function_name,
-            function_args,
-        })
-    }
-}
-
-impl TransactionContractCall {
-    pub fn to_clarity_contract_id(&self) -> QualifiedContractIdentifier {
-        QualifiedContractIdentifier::new(
-            StandardPrincipalData::from(self.address.clone()),
-            self.contract_name.clone(),
-        )
-    }
-}
-
-impl fmt::Display for TransactionContractCall {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let formatted_args = self
-            .function_args
-            .iter()
-            .map(|v| format!("{}", v))
-            .collect::<Vec<String>>()
-            .join(", ");
-        write!(
-            f,
-            "{}.{}::{}({})",
-            self.address, self.contract_name, self.function_name, formatted_args
-        )
-    }
-}
-
-impl StacksMessageCodec for TransactionSmartContract {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        write_next(fd, &self.name)?;
-        write_next(fd, &self.code_body)?;
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<TransactionSmartContract, codec_error> {
-        let name: ContractName = read_next(fd)?;
-        let code_body: StacksString = read_next(fd)?;
-        Ok(TransactionSmartContract { name, code_body })
-    }
-}
 
 fn ClarityVersion_consensus_serialize<W: Write>(
     version: &ClarityVersion,
@@ -363,161 +289,6 @@ impl TransactionPayload {
             )),
             (_, _) => None,
         }
-    }
-}
-
-impl StacksMessageCodec for AssetInfo {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        write_next(fd, &self.contract_address)?;
-        write_next(fd, &self.contract_name)?;
-        write_next(fd, &self.asset_name)?;
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<AssetInfo, codec_error> {
-        let contract_address: StacksAddress = read_next(fd)?;
-        let contract_name: ContractName = read_next(fd)?;
-        let asset_name: ClarityName = read_next(fd)?;
-        Ok(AssetInfo {
-            contract_address,
-            contract_name,
-            asset_name,
-        })
-    }
-}
-
-impl StacksMessageCodec for PostConditionPrincipal {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        match *self {
-            PostConditionPrincipal::Origin => {
-                write_next(fd, &(PostConditionPrincipalID::Origin as u8))?;
-            }
-            PostConditionPrincipal::Standard(ref address) => {
-                write_next(fd, &(PostConditionPrincipalID::Standard as u8))?;
-                write_next(fd, address)?;
-            }
-            PostConditionPrincipal::Contract(ref address, ref contract_name) => {
-                write_next(fd, &(PostConditionPrincipalID::Contract as u8))?;
-                write_next(fd, address)?;
-                write_next(fd, contract_name)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<PostConditionPrincipal, codec_error> {
-        let principal_id: u8 = read_next(fd)?;
-        let principal = match principal_id {
-            x if x == PostConditionPrincipalID::Origin as u8 => PostConditionPrincipal::Origin,
-            x if x == PostConditionPrincipalID::Standard as u8 => {
-                let addr: StacksAddress = read_next(fd)?;
-                PostConditionPrincipal::Standard(addr)
-            }
-            x if x == PostConditionPrincipalID::Contract as u8 => {
-                let addr: StacksAddress = read_next(fd)?;
-                let contract_name: ContractName = read_next(fd)?;
-                PostConditionPrincipal::Contract(addr, contract_name)
-            }
-            _ => {
-                return Err(codec_error::DeserializeError(format!(
-                    "Failed to parse transaction: unknown post condition principal ID {}",
-                    principal_id
-                )));
-            }
-        };
-        Ok(principal)
-    }
-}
-
-impl StacksMessageCodec for TransactionPostCondition {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        match *self {
-            TransactionPostCondition::STX(ref principal, ref fungible_condition, ref amount) => {
-                write_next(fd, &(AssetInfoID::STX as u8))?;
-                write_next(fd, principal)?;
-                write_next(fd, &(*fungible_condition as u8))?;
-                write_next(fd, amount)?;
-            }
-            TransactionPostCondition::Fungible(
-                ref principal,
-                ref asset_info,
-                ref fungible_condition,
-                ref amount,
-            ) => {
-                write_next(fd, &(AssetInfoID::FungibleAsset as u8))?;
-                write_next(fd, principal)?;
-                write_next(fd, asset_info)?;
-                write_next(fd, &(*fungible_condition as u8))?;
-                write_next(fd, amount)?;
-            }
-            TransactionPostCondition::Nonfungible(
-                ref principal,
-                ref asset_info,
-                ref asset_value,
-                ref nonfungible_condition,
-            ) => {
-                write_next(fd, &(AssetInfoID::NonfungibleAsset as u8))?;
-                write_next(fd, principal)?;
-                write_next(fd, asset_info)?;
-                write_next(fd, asset_value)?;
-                write_next(fd, &(*nonfungible_condition as u8))?;
-            }
-        };
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<TransactionPostCondition, codec_error> {
-        let asset_info_id: u8 = read_next(fd)?;
-        let postcond = match asset_info_id {
-            x if x == AssetInfoID::STX as u8 => {
-                let principal: PostConditionPrincipal = read_next(fd)?;
-                let condition_u8: u8 = read_next(fd)?;
-                let amount: u64 = read_next(fd)?;
-
-                let condition_code = FungibleConditionCode::from_u8(condition_u8).ok_or(
-                    codec_error::DeserializeError(format!(
-                    "Failed to parse transaction: Failed to parse STX fungible condition code {}",
-                    condition_u8
-                )),
-                )?;
-
-                TransactionPostCondition::STX(principal, condition_code, amount)
-            }
-            x if x == AssetInfoID::FungibleAsset as u8 => {
-                let principal: PostConditionPrincipal = read_next(fd)?;
-                let asset: AssetInfo = read_next(fd)?;
-                let condition_u8: u8 = read_next(fd)?;
-                let amount: u64 = read_next(fd)?;
-
-                let condition_code = FungibleConditionCode::from_u8(condition_u8).ok_or(
-                    codec_error::DeserializeError(format!(
-                    "Failed to parse transaction: Failed to parse FungibleAsset condition code {}",
-                    condition_u8
-                )),
-                )?;
-
-                TransactionPostCondition::Fungible(principal, asset, condition_code, amount)
-            }
-            x if x == AssetInfoID::NonfungibleAsset as u8 => {
-                let principal: PostConditionPrincipal = read_next(fd)?;
-                let asset: AssetInfo = read_next(fd)?;
-                let asset_value: Value = read_next(fd)?;
-                let condition_u8: u8 = read_next(fd)?;
-
-                let condition_code = NonfungibleConditionCode::from_u8(condition_u8)
-                    .ok_or(codec_error::DeserializeError(format!("Failed to parse transaction: Failed to parse NonfungibleAsset condition code {}", condition_u8)))?;
-
-                TransactionPostCondition::Nonfungible(principal, asset, asset_value, condition_code)
-            }
-            _ => {
-                return Err(codec_error::DeserializeError(format!(
-                    "Failed to aprse transaction: unknown asset info ID {}",
-                    asset_info_id
-                )));
-            }
-        };
-
-        Ok(postcond)
     }
 }
